everything seems to work

+++ models/src/main/java/com/vet24/models/dto/pet/clinicalexamination/ClinicalExaminationDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class ClinicalExaminationDto {
     Long id;
-    Integer weight;
+    Double weight;
     Boolean isCanMove;
     String text;
